+++ ai/src/main/java/com/owl_express/ai/application/dtos/request/MessageCreateRequestDto.java
@@ -1,4 +1,4 @@
-package com.owl_express.ai.application.dtos;
+package com.owl_express.ai.application.dtos.request;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
